+++ utils/setup_base_db.py
@@ -33,8 +33,9 @@ from urllib.request import urljoin, urlretrieve
 # to be removed in favor of a local version
 from Bio.KEGG import REST
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from webapp.lib import db_utils  # noqa
+zdbdir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+sys.path.append(os.path.join(zdbdir, "webapp"))
+from lib import db_utils  # noqa
 
 # from REST documentation, can get a max of 10 queries
 # in kegg_get
@@ -349,8 +350,7 @@ def setup_cog(db, cog_dir):
     cog_names_file.reconfigure(encoding="Latin-1")
 
     cog_ref_data = []
-    for line_no, line in enumerate(cog_names_file):
-        print(line_no, line)
+    for line in cog_names_file:
         tokens = line.strip().split("\t")
         cog_id = int(tokens[0][3:])
         fun = tokens[1].strip()
@@ -359,7 +359,7 @@ def setup_cog(db, cog_dir):
     db.load_cog_ref_data(cog_ref_data)
 
     cog_fun_data = []
-    for line_no, line in enumerate(fun_names_file):
+    for line in fun_names_file:
         function, random_str, description = line.split("\t")
         cog_fun_data.append((function.strip(), description.strip()))
     db.load_cog_fun_data(cog_fun_data)