+++ webapp/views/utils.py
@@ -30,6 +30,9 @@ title2page = {
         'extract_orthogroup', 'heatmap_orthogroup', 'index_comp_orthogroup',
         'orthogroup_comparison', 'pan_genome_orthogroup', 'plot_heatmap_orthogroup',
         'venn_orthogroup'],
+    'Comparisons: Virulence Factors': [
+        'entry_list_vf', 'index_comp_vf', 'extract_vf', 'venn_vf',
+        'vf_comparison', 'plot_heatmap_vf', 'pan_genome_vf'],
     'Genome alignments: Circos plot': ['circos'],
     'Genome alignments: Plot region': ['plot_region'],
     'Genomes: table of contents': ['extract_contigs', 'genomes_intro'],
@@ -43,6 +46,7 @@ title2page = {
     'Pfam domain': ['fam_pfam'],
     'Phylogeny': ['phylogeny_intro'],
     'Table of content: Genomes': ['genomes'],
+    'Virulence Factor Gene': ["fam_vf"],
 }
 
 page2title = {}
@@ -185,3 +189,7 @@ def format_ko_modules(hsh_modules, ko):
     if len(modules) == 0:
         return "-"
     return "<br>".join([format_ko_module(i, d) for i, d in modules])
+
+
+def format_refseqid_to_ncbi(seqid):
+    return f"<a href=\"http://www.ncbi.nlm.nih.gov/protein/{seqid}\">{seqid}</a>"